Using string as uint64_t.

+++ src/Job/CasJob.hpp
@@ -8,12 +8,12 @@ namespace MemcachedNative {
 
 class CasJob : public JobBase {
 public:
-	explicit CasJob(MemcachedNative::Client*& client_, Callback* callback_, char* key_, char* value_, time_t expirationTime_, uint32_t cas_)
+	explicit CasJob(MemcachedNative::Client*& client_, Callback* callback_, char* key_, char* value_, time_t expirationTime_, uint64_t cas_)
 		: JobBase(client_, callback_), key(key_), value(value_), cas(cas_), expirationTime(expirationTime_) { }
 
 	virtual void execute(memcached_st* memcached) {
-		printf("%s %s (%zu): %s (%zu) %u\n", "CasJob execute", key, strlen(key), value, strlen(value), cas);
-		rc = memcached_cas(memcached, key, strlen(key), value, strlen(value), expirationTime, (uint32_t)0, (uint64_t) cas);
+		printf("%s %s (%zu): %s (%zu) %" PRIu64 "\n", "CasJob execute", key, strlen(key), value, strlen(value), cas);
+		rc = memcached_cas(memcached, key, strlen(key), value, strlen(value), expirationTime, (uint32_t)0, cas);
 	}
 
 	virtual ~CasJob() {
@@ -29,7 +29,7 @@ public:
 private:
 	char* key;
 	char* value;
-	uint32_t cas;
+	uint64_t cas;
 	time_t expirationTime;
 };
